Moved squad utils and community utils.
