+++ src/engine/core/managers/dialogs/DialogManager.ts
@@ -17,10 +17,11 @@ import {
 } from "@/engine/core/managers/dialogs/dialog_types";
 import { EGameEvent, EventsManager } from "@/engine/core/managers/events";
 import { assert } from "@/engine/core/utils/assertion";
+import { getObjectCommunity } from "@/engine/core/utils/community";
 import { hasAlifeInfo } from "@/engine/core/utils/info_portion";
 import { parseInfoPortions, parseStringsList } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getObjectCommunity, isObjectWounded } from "@/engine/core/utils/object";
+import { isObjectWounded } from "@/engine/core/utils/object";
 import { FALSE, TRUE } from "@/engine/lib/constants/words";
 import {
   ClientObject,
