+++ src/engine/core/managers/drop/DropManager.ts
@@ -6,9 +6,10 @@ import { IItemDropAmountDescriptor } from "@/engine/core/managers/drop/drop_type
 import { Stalker } from "@/engine/core/objects/server/creature/Stalker";
 import { abort } from "@/engine/core/utils/assertion";
 import { isArtefact, isGrenade, isWeapon } from "@/engine/core/utils/class_ids";
+import { getObjectCommunity } from "@/engine/core/utils/community";
 import { parseNumbersList, parseStringsList } from "@/engine/core/utils/ini";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getObjectCommunity, setItemCondition } from "@/engine/core/utils/object";
+import { setItemCondition } from "@/engine/core/utils/object";
 import { isAmmoSection, isExcludedFromLootDropItemSection, isLootableItemSection } from "@/engine/core/utils/section";
 import { spawnItemsForObject } from "@/engine/core/utils/spawn";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
