+++ src/engine/core/utils/object/object_danger.ts
@@ -5,8 +5,8 @@ import { SimulationBoardManager } from "@/engine/core/managers/simulation/Simula
 import { SmartTerrain } from "@/engine/core/objects/server/smart_terrain";
 import { ESmartTerrainStatus } from "@/engine/core/objects/server/smart_terrain/types";
 import { ISchemeCombatIgnoreState } from "@/engine/core/schemes/stalker/combat_ignore";
+import { getObjectCommunity } from "@/engine/core/utils/community";
 import { pickSectionFromCondList } from "@/engine/core/utils/ini";
-import { getObjectCommunity } from "@/engine/core/utils/object/object_get";
 import { isObjectWounded } from "@/engine/core/utils/object/object_state";
 import { isObjectInZone } from "@/engine/core/utils/position";
 import { logicsConfig } from "@/engine/lib/configs/LogicsConfig";
