+++ src/engine/core/utils/object/object_get.ts
@@ -1,86 +1,7 @@
-import { registry } from "@/engine/core/database";
-import type { SmartTerrain } from "@/engine/core/objects/server/smart_terrain";
-import type { Squad } from "@/engine/core/objects/server/squad";
-import { assertDefined } from "@/engine/core/utils/assertion";
-import { isStalker } from "@/engine/core/utils/class_ids";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { communities, TCommunity } from "@/engine/lib/constants/communities";
-import { MAX_U16 } from "@/engine/lib/constants/memory";
-import {
-  AlifeSimulator,
-  AnyGameObject,
-  ClientObject,
-  Optional,
-  ServerCreatureObject,
-  ServerGroupObject,
-  ServerHumanObject,
-  ServerObject,
-  TIndex,
-  TNumberId,
-  Vector,
-} from "@/engine/lib/types";
+import { ClientObject, Optional, TIndex } from "@/engine/lib/types";
 
 const logger: LuaLogger = new LuaLogger($filename);
-
-/**
- * @param object - any game object used by the game engine.
- * @returns tuple of object position details: id, gvi, lvi, position.
- */
-export function getObjectPositioning(object: AnyGameObject): LuaMultiReturn<[TNumberId, TNumberId, TNumberId, Vector]> {
-  if (type(object.id) === "number") {
-    return $multi(
-      (object as ServerObject).id,
-      (object as ServerObject).m_game_vertex_id,
-      (object as ServerObject).m_level_vertex_id,
-      (object as ServerObject).position
-    );
-  } else {
-    return $multi(
-      (object as ClientObject).id(),
-      (object as ClientObject).game_vertex_id(),
-      (object as ClientObject).level_vertex_id(),
-      (object as ClientObject).position()
-    );
-  }
-}
-
-/**
- * Get squad of provided object.
- *
- * @param object - server or client object
- * @return object squad server object or null
- */
-export function getObjectSquad(object: AnyGameObject): Optional<Squad> {
-  // Get for client object.
-  if (type(object.id) === "function") {
-    const serverObject: Optional<ServerCreatureObject> = registry.simulator.object((object as ClientObject).id());
-
-    return !serverObject || serverObject.group_id === MAX_U16
-      ? null
-      : registry.simulator.object<Squad>(serverObject.group_id);
-  } else {
-    // Get for server object.
-    return (object as ServerCreatureObject).group_id === MAX_U16
-      ? null
-      : registry.simulator.object<Squad>((object as ServerCreatureObject).group_id);
-  }
-}
-
-/**
- * Returns community of provided object.
- *
- * @param object - client object or server stalker/group
- * @returns object community
- */
-export function getObjectCommunity(object: ClientObject | ServerHumanObject | ServerGroupObject): TCommunity {
-  if (isStalker(object)) {
-    return type(object.id) === "function"
-      ? (object as unknown as ClientObject).character_community()
-      : (object as ServerHumanObject).community();
-  }
-
-  return communities.monster;
-}
 
 /**
  * Get active weapon slot of an object for animating.
