+++ src/engine/core/utils/squad/index.ts
@@ -1,2 +1,3 @@
+export * from "@/engine/core/utils/squad/squad_get";
 export * from "@/engine/core/utils/squad/squad_actions";
 export * from "@/engine/core/utils/squad/squad_state";
