+++ src/engine/scripts/declarations/dialogs/dialog_manager.ts
@@ -9,9 +9,9 @@ import {
 } from "@/engine/core/managers/dialogs";
 import type { SmartTerrain } from "@/engine/core/objects/server/smart_terrain";
 import { extern } from "@/engine/core/utils/binding";
+import { getObjectCommunity } from "@/engine/core/utils/community";
 import { getNpcSpeaker } from "@/engine/core/utils/dialog";
 import { LuaLogger } from "@/engine/core/utils/logging";
-import { getObjectCommunity } from "@/engine/core/utils/object";
 import { getObjectSmartTerrain } from "@/engine/core/utils/position";
 import { communities, TCommunity } from "@/engine/lib/constants/communities";
 import { TRUE } from "@/engine/lib/constants/words";
